+++ src/util/internals/Hybrid/index.ts
@@ -1,2 +1,0 @@
-export { createHybrid } from './Hybrid';
-export * from './types';