+++ src/util/internals/index.ts
@@ -1,4 +1,4 @@
+export * from './Proxy';
 export { getRandomInt } from './getRandomInt';
-export * from './Hybrid';
 export { ifNaN } from './ifNaN';
 export { removeFromArray } from './removeFromArray';